requirements.txt
discord.py
requests
apscheduler
python-dotenv
notion-client
flask